♻️ Extract fetchMissingCovers to reusable utility, eliminating code duplication between dashboard and browse pages
